+++ services/api/Cargo.toml
@@ -13,9 +13,9 @@ tokio = { version = "1.34.0", features = ["full"] }
 rand = "0.8.5"
 mongodb = "2.1"
 dotenv = "0.15.0"
-serde_json = "1.0.108"
+serde_json = "1.0.120"
+serde =  "1.0.204" 
 futures = "0.3.29"
-serde = "1.0.193"
 tracing = "0.1.40"
 tower = { version = "0.4.13", features = ["timeout"] }
 axiom-rs = "0.9.0"
