--- services/api/src/entities/user.rs
@@ -1,3 +1,0 @@
-pub struct User {
-    _id: String,
-}
