+++ services/api/src/utils.rs
@@ -1,6 +1,5 @@
 pub mod generate_id;
 pub mod get_current_time;
-pub mod get_env;
 pub mod get_header_value;
 pub mod get_reserved_workspace_ids;
 pub mod headers_to_hashmap;
