+++ services/api/src/utils/logger.rs
@@ -1,7 +1,7 @@
-use super::get_env::get_env;
 use crate::utils::get_current_time::iso_format;
 use axiom_rs::Client;
 use serde::Serialize;
+use shared::get_env::get_env;
 use std::{fmt, sync::Arc};
 use time::OffsetDateTime;
 use tokio::{
